+++ packages/bugsnag_flutter/lib/src/client.dart
@@ -523,7 +523,9 @@ class ChannelClient implements BugsnagClient {
     required bool unhandled,
     required bool deliver,
   }) async {
-    final buildID = error.stacktrace.first.codeIdentifier;
+    final buildID = error.stacktrace.isNotEmpty
+        ? error.stacktrace.first.codeIdentifier
+        : null;
     final errorInfo = details?.informationCollector?.call() ?? [];
     final errorContext = details?.context?.toDescription();
     final errorLibrary = details?.library;